sending assets can now be performed on asset list view
